A couple of translation errors or small improvements
bee: is just 'bij', honingbij would be honeybee
bear: beer
chipmunk: is chipmunch a typo or am I wrong here? If so ignore. Wangzakeekhoorn is the correct translation.
Dinosaur: dinosaurus
Giraffe: giraf
Hen: the female chicken is also called hen in Dutch, kip is just general name for chicken.
Hog: zwijn, varken is a pig
Sheep: As everything is written in singolar, I changed this to singolar as well. Schapen would be the plural.
Turkey: the animal is kalkoen. Turkije would be the country, that is the same in English as the animal.